Synthetic_data_generator/requirements.txt
google-generativeai
sentence-transformers
numpy
torch
tqdm